+++ crates/usvg-parser/src/image.rs
@@ -250,6 +250,7 @@ pub(crate) fn load_sub_svg(data: &[u8], opt: &Options) -> Option<ImageKind> {
             return None;
         }
     };
+    tree.calculate_abs_transforms();
     tree.calculate_bounding_boxes();
 
     Some(ImageKind::SVG(tree))